requirements.txt
discord.py
aiohttp
asyncio
colorama
PyNaCl